backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
click==8.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
resend==2.1.0
python-jose[cryptography]==3.3.0
